requirements.txt
cytoolz==0.11.0
matplotlib==3.5.3
numpy==1.21.5
pandas==1.3.5
pyro_ppl==1.8.4
scikit_learn==1.0.2
scipy==1.7.3
torch==1.12.0+cu116
torch_geometric==2.1.0.post1
tqdm==4.64.1
